move domain helpers functions to separate dir
